+++ CMakeLists.txt
@@ -57,7 +57,7 @@ if (BUILD_EMULATOR)
             src/Interrupts/interrupt_around_function.cpp
             src/Interrupts/interrupt_remote_call.cpp
             src/Interrupts/interrupt_hook_on_addr.cpp
-            src/Interrupts/interrupt_monitor_event.cpp
+            src/Interrupts/interrupt_hook_on_event.cpp
             src/Interrupts/interrupt_response.cpp
             src/Instrumentation/logical_clock.cpp
             src/Instrumentation/hook.cpp)
@@ -118,7 +118,7 @@ if (BUILD_UNITTEST)
             src/Interrupts/interrupt_around_function.cpp
             src/Interrupts/interrupt_remote_call.cpp
             src/Interrupts/interrupt_hook_on_addr.cpp
-            src/Interrupts/interrupt_monitor_event.cpp
+            src/Interrupts/interrupt_hook_on_event.cpp
             src/Interrupts/interrupt_response.cpp
             src/Instrumentation/logical_clock.cpp
             src/Instrumentation/hook.cpp)
