+++ platforms/ESP-IDF/CMakeLists.txt
@@ -18,7 +18,7 @@ set(SOURCE_FILES
         ../../src/Interrupts/interrupt_around_function.cpp
         ../../src/Interrupts/interrupt_remote_call.cpp
         ../../src/Interrupts/interrupt_hook_on_addr.cpp
-        ../../src/Interrupts/interrupt_monitor_event.cpp
+        ../../src/Interrupts/interrupt_hook_on_event.cpp
         ../../src/Interrupts/interrupt_response.cpp
         ../../src/Instrumentation/logical_clock.cpp
         ../../src/Instrumentation/hook.cpp
